server/requirements.txt
Flask
python-dotenv
google-generativeai
Flask-CORS
gunicorn